requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
databento==0.64.0
databento-dbn==0.42.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.9.0
numpy==2.3.4
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
plotly==6.3.1
propcache==0.4.1
protobuf==6.33.0
pyarrow==22.0.0
pydeck==0.9.1
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.2
smmap==5.0.2
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
